requirements.txt
redis>=5.0.0
pynacl>=1.5.0
tree-sitter>=0.22.0
tree-sitter-python>=0.22.0
z3-solver>=4.12.0
coverage>=7.0.0
boto3>=1.34.0
pytest>=7.4.0
mypy>=1.8.0
ruff>=0.5.0
tomli>=2.0.1; python_version < "3.11"

